% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sterr.R
\name{sterr}
\alias{sterr}
\title{Calculate the standard error}
\usage{
sterr(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector for which the standard error will be calculated
from.}

\item{na.rm}{If *NA* values are not an issue, then *FALSE*,
otherwise, *TRUE*.}
}
\description{
The \code{'sterr()'} function calculates the standard error of a variable,
taking into account \code{NA} values, either to keep or remove them
depending on the user input to the \code{na.rm} argument.
}
\examples{
x <- rnorm(10, 5, 0.25)
sterr(x)
}
\keyword{err}
\keyword{se}
\keyword{std}
