% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePackage.R
\name{usePackage}
\alias{usePackage}
\title{Package installing and loading function}
\usage{
usePackage(p)
}
\arguments{
\item{p}{A package name as a string in quotes.}
}
\description{
The usePackage function takes a single command, the name of a package, and will install it, if not already installed, and then load it.
}
\examples{
usePackage("ggplot2")
}
\keyword{pack}
\keyword{use}
