% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usePackages.R
\name{usePackages}
\alias{usePackages}
\title{Package (multiple) installing and loading function}
\usage{
usePackages(..., packlist = NULL)
}
\arguments{
\item{packlist}{A list of package names as a vector of quoted strings.}
}
\description{
The usePackages function takes a single command, a list of package names, and will install them, if not already installed, and then load them.
}
\examples{
usePackage(c("ggplot2", "dplyr"))
}
\keyword{pack}
\keyword{use}
